LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/client-navigate.https-expected.txt



Harness Error (TIMEOUT), message = null

FAIL Frame location should update on successful navigation assert_unreached: unexpected rejection: assert_equals: expected "https://localhost:9443/service-workers/service-worker/resources/client-navigated-frame.html" but got "" Reached unreachable code
FAIL Frame location should not be accessible after redirect assert_unreached: unexpected rejection: assert_throws: function "function () { return frame.contentWindow.location.href }" did not throw Reached unreachable code
FAIL Frame location should not be accessible after cross-origin navigation assert_unreached: unexpected rejection: assert_throws: function "function () { return frame.contentWindow.location.href }" did not throw Reached unreachable code
PASS Frame location should not update on failed about:blank navigation 
PASS Frame location should not update on failed mixed-content navigation 
